requirements-dev.txt
-r requirements.txt

black==22.10.0
faker==15.3.1
flake8==5.0.4
flask-testing==0.8.1
isort==5.10.1
mypy==0.982
pre-commit==2.20.0
pytest==7.2.0
pytest-cov==4.0.0
types-sqlalchemy==1.4.53.3
